LayoutTests/platform/ios-simulator/fast/events/scale-and-scroll-iframe-body-expected.txt


FAIL frame.contentDocument.body.scrollTop should be 100. Was 0.
PASS frame.contentDocument.body.scrollLeft is 100
PASS frame.contentWindow.scrollX is 100
FAIL frame.contentWindow.scrollY should be 100. Was 0.
